added Payment Model

+++ newShopApI/src/main/java/com/example/newShopApI/service/OrderService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.example.newShopApI.exception.MethodNotAllowedException;
 import com.example.newShopApI.exception.ResourceNotFoundException;
 import com.example.newShopApI.model.Order;
 import com.example.newShopApI.model.Product;
@@ -32,26 +33,26 @@ public class OrderService {
 	public Order createNewOrderService(ArrayList<Long> ids, Long userId) {
 		double totalPrice = 0.0;
 		List<ProductVarient> productVarients = new ArrayList<ProductVarient>();
-		for(Long id : ids) {
+		for (Long id : ids) {
 			ProductVarient productVarient = productVarientService.findProductVarientByIdService(id);
 			productVarient.setStock(productVarient.getStock() - 1);
 			productVarientService.updateProductVarientService(id, productVarient);
 			productVarients.add(productVarient);
 			totalPrice = totalPrice + productVarient.getPrice();
 		}
-		
+
 		Order order = new Order();
 		order.setProductVarients(productVarients);
 		order.setUser(userService.findUserById(userId));
 		order.setTotalPrice(totalPrice);
-		
+
 		return orderRepository.save(order);
 	}
 
 	public Order updateOrderService(Long orderId, Order orderDetails) {
 		Order order = orderRepository.findById(orderId)
 				.orElseThrow(() -> new ResourceNotFoundException("Order", "id", orderId));
-		List <ProductVarient> OldproductVarients = order.getProductVarients();
+		List<ProductVarient> OldproductVarients = order.getProductVarients();
 		for (ProductVarient prodVar : OldproductVarients) {
 			prodVar.setStock(prodVar.getStock() + 1);
 			Product product = prodVar.getProduct();
@@ -72,16 +73,22 @@ public class OrderService {
 			totalPrice = totalPrice + productVarient.getPrice();
 		}
 		order.setTotalPrice(totalPrice);
-		
+
 		return orderRepository.save(order);
 	}
-	
+
 	public ResponseEntity<?> deleteOrderService(Long orderId) {
 		Order order = orderRepository.findById(orderId)
 				.orElseThrow(() -> new ResourceNotFoundException("Order", "id", orderId));
-		
+
+		if (order.getPayment() != null)
+			throw new MethodNotAllowedException("Order", "id", orderId);
 		orderRepository.delete(order);
-		
+
 		return ResponseEntity.ok().build();
+	}
+
+	public Order findOrderById(Long id) {
+		return orderRepository.findById(id).orElseThrow(() -> new MethodNotAllowedException("Order", "id", id));
 	}
 }
